Introduce AssistedInject
The ViewModel looks great now: no duplication and the factory is now an interface — AssistedInject generates the implementation for us. The downside is the AssistedInjectModule we have to define, but that’s just a one time step.

The AssistedInjectModule needs to be installed in a component of our choice. Since we’ll most likely never scope bindings from this module anyway (why scope the factory of a ViewModel, if the ViewModel already has its own scope?), it’s fine to stick with the ApplicationComponent.

We also have to unfortunately set a compiler option so Hilt doesn’t complain about the generated assisted inject module not having an @InstallIn annotation, but thankfully this got improved on the latest release of AssistedInject, which the next commit will cover.

# Conflicts:
#	app/src/main/java/com/tfcporciuncula/hiltassistedinjection/OutputViewModel.kt