Introduce a factory
Here we’re making things more safe at the cost of verbosity. The more parameters the ViewModel has, the more verbosity you end up with since you have to repeat them all in the factory. Also, injecting the ViewModel is more painful now as we need to first inject the factory, and then ask for the ViewModel.