Add SavedStateHandle to show how it’d look like
Now that we have AssistedInject, we can’t rely on Hilt’s way of injecting the SavedStateHandle anymore and we have to do it ourselves with AssistedInject. For that, we’ll need to update our extension so that it uses the AbstractSavedStateViewModelFactory so we can access the SavedStateHandle and pass it to our factory.